PurchaseCount, ShowNumbersToggle, Tickets로 컴포넌트 분리
