+++ src/js/PurchaseCount.js
@@ -1,0 +1,11 @@
+class PurchaseCount {
+  constructor() {
+    this.$purchaseCount = document.querySelector("#purchase-count");
+  }
+
+  render({ purchaseCount }) {
+    this.$purchaseCount.innerHTML = purchaseCount;
+  }
+}
+
+export default PurchaseCount;
